Adds inventory checking for App and Monitor servers
The `manage-groups` role dynamically assigns the "app" and "mon"
production hosts to the relevant hostgroups early on in the playbook. If
by chance the Admin has changed the default hostnames for "app" and
"mon" to something that doesn't start with those strings, the playbook
will exit and display an informative message.

The "app" and "mon" naming convention is not new; the playbooks have
used it for a long time now. This change simply improves error-checking
to ensure a smooth provisioning process.